requirements.txt
streamlit
audio_recorder_streamlit
SpeechRecognition
gtts
python-dotenv
openai